requirements.txt
appdirs==1.4.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cycler==0.10.0
distlib==0.3.0
Django==4.0.3
django-filter==22.1
django-parler==2.3
django-rosetta==0.9.8
django-widget-tweaks==1.4.12
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.2.0
kiwisolver==1.3.1
matplotlib==3.3.3
mccabe==0.6.1
nltk==3.8
numpy==1.24.1
pandas==1.5.2
Pillow==8.0.1
pipreqs==0.4.11
polib==1.1.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.6
python-gettext==4.1
pytz==2022.7
regex==2022.10.31
requests==2.27.1
rfc3986==1.5.0
rosetta==0.3
scipy==1.9.1
six==1.14.0
sniffio==1.3.0
sorl-thumbnail==12.9.0
sqlparse==0.4.2
tqdm==4.64.1
translate-po==1.0.13
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.0.17
yarg==0.1.9
